Removes deprecated files and moves to Parcel-based compilation
