--- .surgeignore
@@ -1,5 +1,0 @@
-.zip
-bower_components
-CNAME
-Makefile
-node_modules
