+++ README.md
@@ -72,11 +72,11 @@ $ bundle install
 
 #### Install Node Dependencies
 
-You will need Node v10.0 or greater to compile frontend assets. We're using [Dart Sass](https://github.com/sass/dart-sass) to compile Scss, which needs to be installed globally.
+You will need Node v10.0 or greater to compile frontend assets. We're using [Parcel JS](https://github.com/sass/dart-sass) to compile Scss and JavaScript. Parcel JS needs to be installed globally.
 
 ```bash
-$ npm install -g sass
-# Installs Dart Sass globally
+$ npm install -g parcel-bundler
+# Installs Parcel JS globally
 ```
 
 Next you'll need to install all the other JS dependencies.
@@ -91,13 +91,6 @@ $ npm install .
 ```bash
 $ bundle exec jekyll serve
 # => Now browse to http://localhost:4000
-```
-
-#### Compile CSS
-
-```bash
-$ npm run build-css
-# Compiles the _scss/styles.scss file into static/css/styles.css
 ```
 
 #### Run html-proofer to find broken links and accessibility issues
