+++ _config.yml
@@ -26,6 +26,7 @@ exclude:
   - .editorconfig
   - .gitignore
   - .surgeignore
+  - .cache
   - bower_components
   - CNAME
   - CODE_OF_CONDUCT.md
