+++ _layouts/base.html
@@ -11,7 +11,7 @@
 {% include youtube-embed-head.html id=youtube_id %}
 {% endif %}
     <link href="https://fonts.googleapis.com/css?family=Source+Sans+Pro:400,400i,700,700i|Signika:600,700" rel="stylesheet">
-    <link rel="stylesheet" href="/static/css/styles.css">
+    <link rel="stylesheet" href="/static/main.css">
 
     <!-- favicon
     ================================================== -->
@@ -44,12 +44,7 @@
 
     <!-- Loading JavaScripts
     ================================================== -->
-    <script src="//ajax.googleapis.com/ajax/libs/jquery/2.2.4/jquery.min.js"></script>
-    <script>window.jQuery || document.write('<script src="/static/js/jquery.min.js"><\/script>')</script>
-    <script src="/static/js/what-input.min.js"></script>
-    <script src="/static/js/foundation.min.js"></script>
-    <script src="/static/js/headroom.min.js"></script>
-    <script src="/static/js/main.js"></script>
+    <script src="/static/main.js"></script>
     {% include analytics.html %}
   </body>
 </html>
