+++ _styleguide/template.hbs
@@ -11,14 +11,13 @@ sitemap: false
     <title>Style Guide</title>
 
     <link href="https://fonts.googleapis.com/css?family=Source+Sans+Pro:400,400i,700,700i|Signika:600,700" rel="stylesheet">
-    <link rel="stylesheet" href="/static/css/styles.css">
+    <link rel="stylesheet" href="/static/main.css">
 
     <!-- favicon
     ================================================== -->
     <link href="/static/icons/favicon.ico" rel=icon type="image/x-icon">
     <link href="/static/icons/favicon.png" rel=icon type="image/png">
 
-    <link rel="stylesheet" href="/static/css/styles.css">
     <link rel="stylesheet" href="//cdnjs.cloudflare.com/ajax/libs/highlight.js/8.4/styles/default.min.css">
   </head>
   <body>
@@ -71,11 +70,6 @@ sitemap: false
 
     <!-- Loading JavaScripts
     ================================================== -->
-    <script src="//ajax.googleapis.com/ajax/libs/jquery/2.2.4/jquery.min.js"></script>
-    <script>window.jQuery || document.write('<script src="/static/js/jquery.min.js"><\/script>')</script>
-    <script src="/static/js/what-input.min.js"></script>
-    <script src="/static/js/foundation.min.js"></script>
-    <script src="/static/js/headroom.min.js"></script>
-    <script src="/static/js/main.js"></script>
+    <script src="/static/main.js"></script>
   </body>
 </html>
