+++ package.json
@@ -3,12 +3,15 @@
   "version": "1.0.0",
   "description": "DjangoCon US Website",
   "devDependencies": {
-    "all-contributors-cli": "^4.11.2"
+    "all-contributors-cli": "^4.11.2",
+    "cssnano": "^4.1.8",
+    "sass": "^1.17.0"
   },
   "scripts": {
     "add": "all-contributors add",
     "generate": "all-contributors generate",
-    "build-css": "sass --style=compressed _scss/styles.scss static/css/styles.css -q --no-source-map"
+    "build": "parcel build main.js -d static/ --no-source-maps",
+    "build-dev": "parcel build main.js -d static/ --no-minify"
   },
   "repository": {
     "type": "git",
@@ -20,6 +23,9 @@
   "dependencies": {
     "foundation-sites": "^6.5.3",
     "hamburgers": "^1.1.3",
-    "motion-ui": "^2.0.3"
+    "headroom.js": "^0.9.4",
+    "jquery": "^3.3.1",
+    "motion-ui": "^2.0.3",
+    "what-input": "^5.1.3"
   }
 }
